refactor(src/bridge): reorganize into source-agnostic connectors and hub
Split the Discord-baked one-way mirror into a source-agnostic core, a Discord
connector, and a Linear hub connector, making way for a future GitHub
Discussions bridge without changing behavior or the on-the-wire format.

- core/: canonical model, Source/Target capability interfaces, orchestrating
  Mirror, reconciler, references, rate-limit retry.
- discord/: DiscordConnector (listeners + backfill + announce) and model mapping.
- linear/: hub split into client/issues/comments/reactions/labels/emojis/
  attachments/state/assets behind LinearConnector.
- Add src/bridge/ARCHITECTURE.md documenting the source-authoritative,
  Linear-as-relay-hub reconciliation model and the planned bidirectional path.
- Delete the redundant one-shot sync script (superseded by startup backfill).
