--- src/bridge/linear/api.ts
@@ -1,671 +1,0 @@
-import {
-  LinearClient,
-  type Attachment,
-  type Comment,
-  type Issue,
-} from "@linear/sdk";
-import { IssueRelationType } from "@linear/sdk";
-
-import { config } from "@lib/config.js";
-
-// Validated bridge credentials. Present whenever the bridge is enabled.
-function bridgeConfig(): {
-  appToken: string;
-  userToken: string;
-  teamId: string;
-} {
-  const { appToken, userToken, teamId } = config.linearBridge;
-  if (!appToken || !userToken || !teamId) {
-    throw new Error(
-      "linearBridge is enabled but appToken/userToken/teamId are missing",
-    );
-  }
-  return { appToken, userToken, teamId };
-}
-
-let appClient: LinearClient | undefined;
-let userClient: LinearClient | undefined;
-
-// App-actor client. Issues, comments and reactions run here so they are
-// attributed to the external Discord author (OAuth tokens use accessToken).
-function linear(): LinearClient {
-  if (!appClient) {
-    appClient = new LinearClient({ accessToken: bridgeConfig().appToken });
-  }
-  return appClient;
-}
-
-// Personal-key client for writes the app actor cannot make: creating custom
-// emojis and labels.
-function linearUser(): LinearClient {
-  if (!userClient) {
-    userClient = new LinearClient({ apiKey: bridgeConfig().userToken });
-  }
-  return userClient;
-}
-
-// Extracts a readable message from a Linear SDK error, whose default string
-// form is unhelpful ("[object Object]").
-export function linearError(err: unknown): string {
-  const e = err as { errors?: { message?: string }[]; message?: string };
-  return (
-    e?.errors
-      ?.map((x) => x.message)
-      .filter(Boolean)
-      .join("; ") ||
-    e?.message ||
-    String(err)
-  );
-}
-
-// Whether an error is a Linear rate-limit rejection, so a bulk import can wait
-// and retry rather than abort.
-export function isRateLimited(err: unknown): boolean {
-  const e = err as {
-    type?: string;
-    status?: number;
-    errors?: { extensions?: { type?: string } }[];
-  };
-  return (
-    e?.type === "Ratelimited" ||
-    e?.status === 429 ||
-    (e?.errors?.some((x) => x.extensions?.type === "Ratelimited") ?? false)
-  );
-}
-
-// Metadata stored on the Discord attachment of a mirrored issue.
-export interface ThreadAttachmentFields {
-  url: string;
-  title: string;
-  subtitle: string;
-  metadata: Record<string, unknown>;
-}
-
-// Returns the first attachment on the URL whose issue lives in the configured
-// team. Attachments match across the whole workspace, so scoping to the team
-// keeps lookups from touching issues in unrelated Linear teams.
-async function attachmentInTeam(
-  url: string,
-): Promise<{ attachment: Attachment; issue: Issue } | null> {
-  const { teamId } = bridgeConfig();
-  const attachments = await linear().attachmentsForURL(url);
-  for (const attachment of attachments.nodes) {
-    const issue = await attachment.issue;
-    if (!issue) continue;
-    const team = await issue.team;
-    if (team?.id === teamId) return { attachment, issue };
-  }
-  return null;
-}
-
-// Finds the issue mapped to a thread via its URL attachment, returning the
-// issue and attachment ids.
-export async function findThreadMapping(
-  url: string,
-): Promise<{ issueId: string; attachmentId: string } | null> {
-  const match = await attachmentInTeam(url);
-  if (!match) return null;
-  return { issueId: match.issue.id, attachmentId: match.attachment.id };
-}
-
-// Creates an issue in the configured team and returns its id.
-export async function createIssue(input: {
-  title: string;
-  description: string;
-  author?: { name: string; iconUrl?: string };
-  createdAt?: Date;
-}): Promise<string> {
-  const payload = await linear().createIssue({
-    teamId: bridgeConfig().teamId,
-    projectId: config.linearBridge.projectId,
-    title: input.title,
-    description: input.description,
-    createdAt: input.createdAt,
-    // Attributes the issue to an external Discord author under app-actor auth;
-    // ignored fields are safe to omit for personal keys (author is undefined).
-    createAsUser: input.author?.name,
-    displayIconUrl: input.author?.iconUrl,
-  });
-
-  const issue = await payload.issue;
-  if (!issue) throw new Error("Linear did not return the created issue");
-  console.log(`[bridge] created issue ${issue.identifier} "${input.title}"`);
-  return issue.id;
-}
-
-// Creates the Discord attachment on a freshly created issue.
-export async function createThreadAttachment(
-  issueId: string,
-  fields: ThreadAttachmentFields,
-): Promise<void> {
-  await linear().createAttachment({ issueId, ...fields });
-  console.log(`[bridge] created attachment on ${issueId} -> ${fields.url}`);
-}
-
-// Updates the issue's Discord attachment in place, or creates it if missing.
-export async function upsertThreadAttachment(
-  issueId: string,
-  fields: ThreadAttachmentFields,
-): Promise<void> {
-  const mapping = await findThreadMapping(fields.url);
-  if (!mapping) {
-    await createThreadAttachment(issueId, fields);
-    return;
-  }
-
-  await linear().updateAttachment(mapping.attachmentId, {
-    title: fields.title,
-    subtitle: fields.subtitle,
-    metadata: fields.metadata,
-  });
-  console.log(`[bridge] updated attachment on ${issueId}`);
-}
-
-// Invisible marker (an unused markdown reference-link definition) appended to
-// mirrored comments so a later Discord edit/delete can locate the right
-// comment. It renders as nothing in Linear but round-trips in the raw body.
-const MSG_MARKER = "discord-msg";
-
-function withMarker(body: string, messageId: string): string {
-  return `${body}\n\n[${MSG_MARKER}]: ${messageId}`;
-}
-
-function markerMessageId(body: string): string | null {
-  return body.match(/^\[discord-msg\]:\s*(\S+)/m)?.[1] ?? null;
-}
-
-export async function addComment(
-  issueId: string,
-  body: string,
-  author?: { name: string; iconUrl?: string },
-  messageId?: string,
-  parentId?: string,
-  createdAt?: Date,
-): Promise<void> {
-  const input = {
-    issueId,
-    body: messageId ? withMarker(body, messageId) : body,
-    createdAt,
-    // Attributes the comment to an external Discord author. Requires OAuth
-    // app-actor auth; Linear rejects these fields for personal API keys, so the
-    // caller only supplies an author when that mode is configured.
-    createAsUser: author?.name,
-    displayIconUrl: author?.iconUrl,
-  };
-
-  console.log(
-    `[bridge] adding comment on ${issueId}` +
-      `${messageId ? ` for msg ${messageId}` : ""}` +
-      `${parentId ? ` (reply to ${parentId})` : ""}`,
-  );
-
-  try {
-    await linear().createComment({ ...input, parentId });
-  } catch (err) {
-    // Linear threads are one level deep; if the parent is itself a reply, fall
-    // back to a top-level comment rather than dropping the message.
-    if (!parentId) throw err;
-    await linear().createComment(input);
-  }
-}
-
-// Returns the Discord message ids already mirrored as comments on the issue,
-// read from the invisible markers, so a backfill can skip them.
-export async function mirroredMessageIds(
-  issueId: string,
-): Promise<Set<string>> {
-  const issue = await linear().issue(issueId);
-  const ids = new Set<string>();
-
-  let page = await issue.comments({ first: 100 });
-  while (true) {
-    for (const comment of page.nodes) {
-      const id = markerMessageId(comment.body);
-      if (id) ids.add(id);
-    }
-    if (!page.pageInfo.hasNextPage) break;
-    page = await issue.comments({
-      first: 100,
-      after: page.pageInfo.endCursor ?? undefined,
-    });
-  }
-  return ids;
-}
-
-// Updates the mirrored comment for a Discord message. Returns false if the
-// message has no mirrored comment.
-export async function editComment(
-  issueId: string,
-  messageId: string,
-  body: string,
-): Promise<boolean> {
-  const commentId = await findCommentByMessage(issueId, messageId);
-  if (!commentId) return false;
-  console.log(`[bridge] editing comment for msg ${messageId} on ${issueId}`);
-  await linear().updateComment(commentId, {
-    body: withMarker(body, messageId),
-  });
-  return true;
-}
-
-// Deletes the mirrored comment for a Discord message. Returns false if the
-// message has no mirrored comment. If the comment has replies, its body is
-// blanked instead of deleted, since Linear removes a comment's replies along
-// with it.
-export async function deleteComment(
-  issueId: string,
-  messageId: string,
-): Promise<boolean> {
-  const node = await findCommentNode(issueId, messageId);
-  if (!node) return false;
-
-  const children = await node.children();
-  if (children.nodes.length > 0) {
-    console.log(
-      `[bridge] tombstoning comment for msg ${messageId} on ${issueId} (has replies)`,
-    );
-    await linear().updateComment(node.id, {
-      body: withMarker("_Message deleted._", messageId),
-    });
-  } else {
-    console.log(`[bridge] deleting comment for msg ${messageId} on ${issueId}`);
-    await linear().deleteComment(node.id);
-  }
-  return true;
-}
-
-// Finds the mirrored comment node for a Discord message id, or null.
-async function findCommentNode(
-  issueId: string,
-  messageId: string,
-): Promise<Comment | null> {
-  const issue = await linear().issue(issueId);
-  const { nodes } = await issue.comments();
-  for (const comment of nodes) {
-    if (markerMessageId(comment.body) === messageId) return comment;
-  }
-  return null;
-}
-
-// Finds the mirrored comment id for a Discord message id, or null.
-export async function findCommentByMessage(
-  issueId: string,
-  messageId: string,
-): Promise<string | null> {
-  return (await findCommentNode(issueId, messageId))?.id ?? null;
-}
-
-// Resolves the comment a Discord reply should attach to: the mirrored comment
-// of the referenced message, collapsed to its thread root since Linear threads
-// are only one level deep. Returns null when the reference isn't mirrored.
-export async function resolveReplyParent(
-  issueId: string,
-  messageId: string,
-): Promise<string | null> {
-  const node = await findCommentNode(issueId, messageId);
-  if (!node) return null;
-  const parent = await node.parent;
-  return parent?.id ?? node.id;
-}
-
-// Replaces an issue's description, used when the opening post is edited.
-export async function setIssueDescription(
-  issueId: string,
-  description: string,
-): Promise<void> {
-  console.log(`[bridge] updating description on ${issueId}`);
-  await linear().updateIssue(issueId, { description });
-}
-
-// Re-hosts a remote file in Linear's storage and returns its permanent asset
-// URL, or null if the upload fails (caller falls back to the source URL).
-export async function uploadFile(
-  sourceUrl: string,
-  filename: string,
-  contentType: string | null,
-): Promise<string | null> {
-  try {
-    const asset = await rehost(
-      sourceUrl,
-      filename,
-      contentType || "application/octet-stream",
-    );
-    console.log(
-      `[bridge] uploaded file ${filename} -> ${asset ? "ok" : "failed"}`,
-    );
-    return asset;
-  } catch {
-    return null;
-  }
-}
-
-// Fetches a remote file and uploads its bytes to Linear storage, returning the
-// permanent asset URL. Linear only accepts asset URLs on its own upload domain,
-// so emojis and attachments must be re-hosted here rather than hotlinked.
-async function rehost(
-  sourceUrl: string,
-  filename: string,
-  type: string,
-): Promise<string | null> {
-  const source = await fetch(sourceUrl);
-  if (!source.ok) return null;
-  const bytes = await source.arrayBuffer();
-
-  const upload = (await linear().fileUpload(type, filename, bytes.byteLength))
-    .uploadFile;
-  if (!upload) return null;
-
-  const headers = new Headers({ "Content-Type": type });
-  for (const { key, value } of upload.headers) headers.set(key, value);
-
-  const put = await fetch(upload.uploadUrl, {
-    method: "PUT",
-    headers,
-    body: bytes,
-  });
-  return put.ok ? upload.assetUrl : null;
-}
-
-// Trashes an issue (recoverable in Linear).
-export async function deleteIssue(issueId: string): Promise<void> {
-  console.log(`[bridge] trashing issue ${issueId}`);
-  await linear().deleteIssue(issueId);
-}
-
-// Reconciles the issue's title and project against the thread, when they drift
-// (e.g. the thread was renamed, or created before the project was configured).
-// One fetch, one update, only when something actually changed.
-export async function reconcileIssue(
-  issueId: string,
-  title: string,
-): Promise<void> {
-  const { projectId } = config.linearBridge;
-  const issue = await linear().issue(issueId);
-
-  const update: { title?: string; projectId?: string } = {};
-  if (issue.title !== title) update.title = title;
-  if (projectId && issue.projectId !== projectId) update.projectId = projectId;
-  if (Object.keys(update).length === 0) return;
-
-  console.log(
-    `[bridge] reconciling issue ${issueId}: ${Object.keys(update).join(", ")}`,
-  );
-  await linear().updateIssue(issueId, update);
-}
-
-// Returns an issue's identifier and URL, e.g. for linking back from Discord.
-export async function getIssueRef(
-  issueId: string,
-): Promise<{ identifier: string; url: string }> {
-  const issue = await linear().issue(issueId);
-  return { identifier: issue.identifier, url: issue.url };
-}
-
-// Returns the workflow state type and name of an issue (e.g. type "started",
-// name "In Progress").
-export async function getIssueState(
-  issueId: string,
-): Promise<{ type: string; name: string } | null> {
-  const issue = await linear().issue(issueId);
-  const state = await issue.state;
-  return state ? { type: state.type, name: state.name } : null;
-}
-
-// Moves an issue to a workflow state of the given type in the team. The started
-// type has several states (In Progress, Blocked, In Review), so pass the state
-// name; without one, the lowest-position state of the type is used.
-export async function setIssueState(
-  issueId: string,
-  type: "completed" | "triage" | "started",
-  preferredName?: string,
-): Promise<void> {
-  const stateId = await findStateId(type, preferredName);
-  if (!stateId) return;
-  console.log(
-    `[bridge] setting issue ${issueId} state -> ${preferredName ?? type}`,
-  );
-  await linear().updateIssue(issueId, { stateId });
-}
-
-const stateIdByType = new Map<string, string>();
-
-// Finds a workflow state of the given type in the team. When preferredName is
-// set, a state with that name wins; otherwise the lowest-position state of the
-// type is used, since Linear does not order the results.
-async function findStateId(
-  type: string,
-  preferredName?: string,
-): Promise<string | null> {
-  const cacheKey = preferredName ? `${type}:${preferredName}` : type;
-  const cached = stateIdByType.get(cacheKey);
-  if (cached) return cached;
-
-  const { teamId } = bridgeConfig();
-  const states = await linear().workflowStates({
-    filter: { team: { id: { eq: teamId } }, type: { eq: type } },
-  });
-
-  const named =
-    preferredName &&
-    states.nodes.find(
-      (s) => s.name.toLowerCase() === preferredName.toLowerCase(),
-    );
-  const byPosition = [...states.nodes].sort((a, b) => a.position - b.position);
-  const id = (named || byPosition[0])?.id ?? null;
-  if (id) stateIdByType.set(cacheKey, id);
-  return id;
-}
-
-// --- Emojis & reactions ---------------------------------------------------
-
-// Names of the workspace's custom emojis, loaded once and updated as we create
-// new ones, so we don't recreate existing emojis or spam duplicate errors.
-let emojiNames: Set<string> | undefined;
-
-async function loadEmojiNames(): Promise<Set<string>> {
-  if (emojiNames) return emojiNames;
-  const names = new Set<string>();
-  let after: string | undefined;
-  do {
-    const page = await linearUser().emojis({ first: 250, after });
-    for (const e of page.nodes) names.add(e.name);
-    after = page.pageInfo.hasNextPage
-      ? (page.pageInfo.endCursor ?? undefined)
-      : undefined;
-  } while (after);
-  emojiNames = names;
-  return names;
-}
-
-// Registers a Discord custom emoji as a workspace emoji named discord-<id> so
-// that :discord-<id>: renders inline. Idempotent: skips emojis that already
-// exist and needs the user token, as the app actor cannot create emojis.
-export async function ensureEmoji(
-  id: string,
-  animated: boolean,
-): Promise<void> {
-  const name = `discord-${id}`;
-  const names = await loadEmojiNames();
-  if (names.has(name)) return;
-
-  const ext = animated ? "gif" : "png";
-  // Linear rejects external image URLs, so re-host the Discord emoji first.
-  const asset = await rehost(
-    `https://cdn.discordapp.com/emojis/${id}.${ext}`,
-    `${name}.${ext}`,
-    animated ? "image/gif" : "image/png",
-  ).catch(() => null);
-  if (!asset) {
-    console.error(`[bridge] ensureEmoji ${name}: upload failed`);
-    return;
-  }
-
-  try {
-    await linearUser().createEmoji({ name, url: asset });
-    names.add(name);
-    console.log(`[bridge] registered emoji ${name}`);
-  } catch (err) {
-    console.error(`[bridge] ensureEmoji ${name} failed:`, linearError(err));
-  }
-}
-
-export type ReactionTarget = { issueId: string } | { commentId: string };
-
-// Reaction ids we created, keyed by target+emoji, so a later removal can delete
-// the exact reaction even for unicode emojis whose stored name differs from the
-// input we sent.
-const reactionIds = new Map<string, string>();
-
-function reactionKey(target: ReactionTarget, emoji: string): string {
-  const scope =
-    "issueId" in target ? `i:${target.issueId}` : `c:${target.commentId}`;
-  return `${scope}|${emoji}`;
-}
-
-export async function addReaction(
-  target: ReactionTarget,
-  emoji: string,
-): Promise<void> {
-  try {
-    const payload = await linear().createReaction({ ...target, emoji });
-    const reaction = await payload.reaction;
-    if (reaction) reactionIds.set(reactionKey(target, emoji), reaction.id);
-    console.log(`[bridge] added reaction ${reactionKey(target, emoji)}`);
-  } catch (err) {
-    console.error(`[bridge] addReaction ${emoji} failed:`, linearError(err));
-  }
-}
-
-export async function removeReaction(
-  target: ReactionTarget,
-  emoji: string,
-): Promise<void> {
-  const key = reactionKey(target, emoji);
-  const id = reactionIds.get(key) ?? (await findReaction(target, emoji));
-  if (!id) return;
-  await linear().deleteReaction(id);
-  reactionIds.delete(key);
-  console.log(`[bridge] removed reaction ${key}`);
-}
-
-// Finds a reaction on the target whose stored emoji matches, used as a fallback
-// when the created id is not cached (e.g. after a restart). Reliable for custom
-// emojis; unicode names are normalized by Linear so may not match.
-async function findReaction(
-  target: ReactionTarget,
-  emoji: string,
-): Promise<string | null> {
-  const reactions =
-    "issueId" in target
-      ? (await linear().issue(target.issueId)).reactions
-      : (await linear().comment({ id: target.commentId })).reactions;
-  return reactions.find((r) => r.emoji === emoji)?.id ?? null;
-}
-
-// --- Labels ---------------------------------------------------------------
-
-// Flat (ungrouped) labels are used rather than a label group because Linear
-// allows only one label per group on an issue, while a help thread can carry
-// several tags. Each label is namespaced by name, e.g. "#help > tag".
-const labelIdByName = new Map<string, string>();
-
-// Finds or creates a team label with the given name, tagging its description
-// with the Discord tag id. Cached by name. Runs on the user token, which owns
-// label management.
-export async function ensureLabel(
-  name: string,
-  tagId: string,
-): Promise<string> {
-  const cached = labelIdByName.get(name);
-  if (cached) return cached;
-
-  const { teamId } = bridgeConfig();
-  const existing = await linearUser().issueLabels({
-    filter: { name: { eq: name }, team: { id: { eq: teamId } } },
-  });
-
-  let id = existing.nodes[0]?.id;
-  if (!id) {
-    const payload = await linearUser().createIssueLabel({
-      name,
-      description: tagId,
-      teamId,
-    });
-    const label = await payload.issueLabel;
-    if (!label) throw new Error("Linear did not return the created label");
-    id = label.id;
-    console.log(`[bridge] created label "${name}"`);
-  }
-
-  labelIdByName.set(name, id);
-  return id;
-}
-
-// Reconciles the issue's namespaced labels to exactly match desiredIds, adding
-// missing ones and removing only stale labels that share the namespace prefix
-// (so unrelated labels are never touched, and labels already absent are never
-// "removed"). Runs on the user token that owns the labels.
-export async function setNamespacedLabels(
-  issueId: string,
-  prefix: string,
-  desiredIds: string[],
-): Promise<void> {
-  const issue = await linearUser().issue(issueId);
-  const current = (await issue.labels()).nodes;
-  const ours = current
-    .filter((l) => l.name.startsWith(prefix))
-    .map((l) => l.id);
-
-  const addedLabelIds = desiredIds.filter((id) => !ours.includes(id));
-  const removedLabelIds = ours.filter((id) => !desiredIds.includes(id));
-  if (addedLabelIds.length === 0 && removedLabelIds.length === 0) return;
-
-  console.log(
-    `[bridge] updating labels on ${issueId}: +${addedLabelIds.length} -${removedLabelIds.length}`,
-  );
-  await linearUser().updateIssue(issueId, { addedLabelIds, removedLabelIds });
-}
-
-// --- Cross-links ----------------------------------------------------------
-
-export interface LinkedIssue {
-  id: string;
-  identifier: string;
-  url: string;
-}
-
-// Finds the Linear issue mapped to a URL via its attachments (a mirrored
-// Discord thread, or a GitHub issue linked through Linear's integration).
-export async function resolveIssueByUrl(
-  url: string,
-): Promise<LinkedIssue | null> {
-  const match = await attachmentInTeam(url);
-  if (!match) return null;
-  const { issue } = match;
-  return { id: issue.id, identifier: issue.identifier, url: issue.url };
-}
-
-// Relation pairs created this session, to avoid duplicate "related" links when
-// the same issue is mentioned more than once.
-const relatedPairs = new Set<string>();
-
-// Marks two issues as related. Idempotent within a session and tolerant of
-// Linear rejecting an existing relation.
-export async function relateIssues(
-  issueId: string,
-  relatedIssueId: string,
-): Promise<void> {
-  const key = [issueId, relatedIssueId].sort().join("|");
-  if (relatedPairs.has(key)) return;
-  relatedPairs.add(key);
-  try {
-    await linear().createIssueRelation({
-      issueId,
-      relatedIssueId,
-      type: IssueRelationType.Related,
-    });
-    console.log(`[bridge] related issues ${key}`);
-  } catch (err) {
-    console.error(`[bridge] relateIssues ${key} failed:`, linearError(err));
-  }
-}
