+++ src/index.ts
@@ -5,7 +5,7 @@ import registerCommandEvents from "./events/commands.js";
 import registerWalkthroughEvents from "./events/walkthrough.js";
 import registerMessageEvents from "./events/messages.js";
 import registerChannelEvents from "./events/channels.js";
-import registerBridgeEvents, { backfillHelpThreads } from "./events/bridge.js";
+import { registerBridge, backfillBridge } from "@bridge/core/bridge.js";
 
 import {
   Client,
@@ -59,7 +59,7 @@ client.once(Events.ClientReady, () => {
   registerWalkthroughEvents(client);
   registerMessageEvents(client);
   registerChannelEvents(client);
-  registerBridgeEvents(client);
+  registerBridge(client);
 
   shufflePresence();
   setInterval(shufflePresence, config.presenceDelay);
@@ -68,7 +68,7 @@ client.once(Events.ClientReady, () => {
     console.error("Failed to catch up on help posts:", err),
   );
 
-  backfillHelpThreads(client).catch((err) =>
+  backfillBridge(client).catch((err) =>
     console.error("Linear bridge: backfill failed:", err),
   );
 });
